pyproject.toml
[tool.poetry]
name = "rd-course-project"
version = "0.1.0"
description = ""
authors = ["DL <dmitriy.luzanovsky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {version = "^2.0.33", extras = ["asyncio"]}
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
gitpython = "^3.1.43"
tqdm = "^4.66.5"
uvicorn = "^0.30.6"
jinja2 = "^3.1.4"
fastapi-pagination = "^0.12.26"
fastapi = {version = "^0.112.2", extras = ["all"]}
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
